+++ packages/builder/plugin-swc/tests/fixtures/react/fast-refresh/expected.js
@@ -1,28 +1,17 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", {
-    value: true
-});
-Object.defineProperty(exports, "App", {
-    enumerable: true,
-    get: function() {
-        return App;
-    }
-});
-var _interopRequireWildcard = require("@swc/helpers/lib/_interop_require_wildcard.js").default;
-var _slicedToArray = require("@swc/helpers/lib/_sliced_to_array.js").default;
-var _jsxDevRuntime = require("react/jsx-dev-runtime");
-var _react = /*#__PURE__*/ _interopRequireWildcard(require("react"));
-var _this = void 0;
+import _sliced_to_array from "@swc/helpers/src/_sliced_to_array.mjs";
+var _this = this;
+import { jsxDEV as _jsxDEV } from "react/jsx-dev-runtime";
 var _s = $RefreshSig$();
-var App = function() {
+import React, { useState } from "react";
+export var App = function() {
     _s();
-    var ref = _slicedToArray((0, _react.useState)(0), 2), count = ref[0], setCount = ref[1];
-    return /*#__PURE__*/ (0, _jsxDevRuntime.jsxDEV)("div", {
+    var ref = _sliced_to_array(useState(0), 2), count = ref[0], setCount = ref[1];
+    return /*#__PURE__*/ _jsxDEV("div", {
         children: [
             "Hello World. ",
             count,
             " ",
-            /*#__PURE__*/ (0, _jsxDevRuntime.jsxDEV)("button", {
+            /*#__PURE__*/ _jsxDEV("button", {
                 onClick: function() {
                     return setCount(function(c) {
                         return c + 1;