+++ packages/solutions/app-tools/src/utils/commands.ts
@@ -1,0 +1,5 @@
+export const getCommand = () => {
+  const args = process.argv.slice(2);
+  const command = args[0];
+  return command;
+};