+++ tests/e2e/webpack-builder/asset-fallback/src/index.js
@@ -1,0 +1,3 @@
+import './foo.xxx';
+
+console.log('1');