+++ packages/cli/core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @modern-js/core
 
+## 1.21.2
+
+### Patch Changes
+
+- Updated dependencies [9d4c0ba]
+  - @modern-js/plugin@1.21.2
+  - @modern-js/node-bundle-require@1.21.2
+  - @modern-js/utils@1.21.2
+
 ## 1.21.1
 
 ### Patch Changes