+++ packages/generator/new-action/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @modern-js/new-action
 
+## 1.21.2
+
+### Patch Changes
+
+- Updated dependencies [7c89bd6]
+  - @modern-js/generator-common@2.4.7
+  - @modern-js/generator-utils@2.4.7
+  - @modern-js/utils@1.21.2
+
 ## 1.21.1
 
 ### Patch Changes