+++ packages/review/eslint-config-app/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @modern-js-app/eslint-config
 
+## 1.21.2
+
+### Patch Changes
+
+- @modern-js/babel-preset-app@1.21.2
+
 ## 1.21.1
 
 ### Patch Changes