+++ packages/server/bff-core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @modern-js/bff-core
 
+## 1.21.2
+
+### Patch Changes
+
+- 9d4c0ba: chore: add types for exports field
+  chore: 为 exports 补充 types 字段
+  - @modern-js/bff-runtime@1.21.2
+  - @modern-js/utils@1.21.2
+
 ## 1.21.1
 
 ### Patch Changes