+++ packages/toolkit/node-bundle-require/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @modern-js/node-bundle-require
 
+## 1.21.2
+
+### Patch Changes
+
+- @modern-js/utils@1.21.2
+
 ## 1.21.1
 
 ### Patch Changes